+++ lib/radfish.rb
@@ -99,9 +99,11 @@ require_relative 'radfish/power_info'
 require_relative 'radfish/boot_info'
 require_relative 'radfish/thermal_info'
 require_relative 'radfish/pci_info'
+require_relative 'radfish/power_distribution_info'
 require_relative 'radfish/controller'
 require_relative 'radfish/volume'
 require_relative 'radfish/client'
+require_relative 'radfish/pdu_adapter'
 
 # Auto-load adapters if available
 begin